requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
PyJWT<=2.9
passlib
gunicorn
psycopg2
stripe
geopandas
pytest
pytest-mock
flask_cors
freezegun
postmarker
fiona
gevent
mailosaur
bs4
geojson
celery
redis